Moved Login/Logout endpoints into their own blueprint under the 'authentication' pkg.  Moved flask functions out of the flask app and into flaskutils.  Moved constants our of flaskapp and into config
