+++ mycodo/mycodo_flask/authentication/__init__.py
@@ -1,0 +1,4 @@
+# coding=utf-8
+"""  """
+
+from . import views
